Added getDate
Added function to get date in yyyymmdd format from the text title of
the XML node. Incorporated this into the getPomodoros method and made
appropriate modifications to the function as well as to the pomodoros
data frame. Also added function to get results for a single month. This
function will be called within the loop to get results for a year.